requirements.txt
alembic>=1.8.0
anyio>=3.6.1
asgiref>=3.5.2
asyncpg>=0.26.0
attrs>=22.1.0
bandit>=1.7.4
bleach>=5.0.1
certifi>=2022.9.14
cffi>=1.15.1
charset-normalizer>=2.1.1
click>=8.1.3
commonmark>=0.9.1
cryptography>=38.0.1
darglint>=1.8.1
dependency-injector>=4.41.0
docutils>=0.19
ecdsa>=0.18.0
eradicate>=2.1.0
fastapi>=0.85.0
fastapi_keycloak>=1.0.7
gitdb>=4.0.9
GitPython>=3.1.27
greenlet>=1.1.3
h11>=0.13.0
idna>=3.4
importlib-metadata>=4.12.0
iniconfig>=1.1.1
itsdangerous>=2.1.2
jaraco.classes>=3.2.2
jeepney>=0.8.0
Jinja2>=3.1.2
keyring>=23.9.1
loguru>=0.7.0
Mako>=1.2.2
MarkupSafe>=2.1.1
mccabe
more-itertools>=8.14.0
nextcloud-api-wrapper>=0.2.3
packaging>=21.3
passlib[bcrypt]==1.7.4
pbr>=5.10.0
pep8-naming>=0.13.2
pkginfo>=1.8.3
pluggy>=1.0.0
psycopg2-binary>=2.9.3
py>=1.11.0
pyasn1>=0.4.8
pycodestyle>=2.8.0
pycparser>=2.21
pydantic==1.9.1
pydocstyle>=6.1.1
pyflakes>=2.4.0
Pygments>=2.13.0
PyJWT>=2.4.0
pyparsing>=3.0.9
pytest>=7.1.3
pytest-asyncio>=0.19.0
python-dotenv>=0.21.0
python-jose>=3.3.0
python-multipart>=0.0.5
PyYAML>=6.0
readme-renderer>=37.1
requests>=2.28.1
requests-toolbelt>=0.9.1
rfc3986>=2.0.0
rich>=12.5.1
rsa>=4.9
SecretStorage>=3.3.3
sentry-sdk>=1.9.0
six>=1.16.0
smmap>=5.0.0
sniffio>=1.3.0
snowballstemmer>=2.2.0
stackprinter
sqladmin>=0.1.12
SQLAlchemy==2.0.22
starlette>=0.20.4
stevedore>=4.0.0
tomli>=2.0.1
twine>=4.0.1
typing_extensions>=4.3.0
urllib3>=1.26.12
uvicorn>=0.18.2
webencodings>=0.5.1
Werkzeug>=2.1.2
WTForms>=3.0.1
zipp>=3.8.1
